Finished video 7.3.
Amended captured widget so that player can now retry or go back to the main menu.

Updated code so that when the player is caught, all sounds stop, except a player caught sound effect which has been imported.  All enemy guards also stop so that the player cannot be "caught" multiple times.